refactor(developer/compilers): splits potentially-common defaults to own file
